requirements.txt
transformers>=4.36.0
torch>=2.1.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
datasets>=2.16.0
streamlit>=1.29.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0
sentencepiece>=0.1.99
protobuf>=3.20.0
